\name{combat_example}
\alias{combat_example}
\title{Simulated MRI data for combat_fit/apply}
\description{Simulated MRI data (FreeSurfer subcortical volumes) to provide an example for \code{\link{combat_fit}} and \code{\link{combat_apply}}.}
\author{Joaquim Radua}
\seealso{\code{\link{combat_fit}} and \code{\link{combat_apply}}}
